requirements.txt
python-telegram-bot[job-queue]==21.4
httpx==0.27.2
PyMySQL==1.1.1
